pyproject.toml
[project]
name = "phalanx"
version = "0.1.0"
description = "an AI-powered cold outreach system for the German market, specifically for identifying potential leads, personalizing outreach emails, and automating the process while maintaining scalability"
authors = [
    {name = "Adebayo Stephen"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openai (>=1.69.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sendgrid (>=6.11.0,<7.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "google-genai (>=1.8.0,<2.0.0)",
    "sib-api-v3-sdk (>=7.6.0,<8.0.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"